General fixes

+++ lib/core/theme/fluent_color_extensions.dart
@@ -2,12 +2,12 @@ import 'package:flutter/material.dart';
 import 'package:fluent_ui/fluent_ui.dart' as fluent;
 
 /// Utility class for converting between Material and Fluent UI colors.
-/// 
+///
 /// This class provides static methods to avoid extension conflicts with
 /// the built-in fluent_ui extensions.
 class FluentColorExtensions {
   /// Converts a Material [Color] to a Fluent UI [AccentColor].
-  /// 
+  ///
   /// This method creates a proper accent color with light and dark variants
   /// based on the original color, suitable for Fluent UI theming.
   static fluent.AccentColor colorToAccentColor(Color color) {
@@ -22,9 +22,9 @@ class FluentColorExtensions {
       'lightest': _adjustBrightness(color, 0.4),
     });
   }
-  
+
   /// Creates a Fluent UI [WidgetStateProperty] color from a Material [Color].
-  /// 
+  ///
   /// This method creates a widget state property with different colors for different states
   /// (normal, pressed, hovered, etc.) based on the original color.
   static fluent.WidgetStateProperty<Color> colorToButtonState(Color color) {
@@ -37,17 +37,15 @@ class FluentColorExtensions {
       return color;
     });
   }
-  
+
   /// Creates a Fluent UI [WidgetStateProperty] for text from a Material [Color].
-  /// 
+  ///
   /// This method creates a widget state property for text with different colors for different states
   /// based on the original color, with appropriate contrast adjustments.
-  static fluent.WidgetStateProperty<Color> colorToTextButtonState(
-    Color color, {
-    bool inverted = false
-  }) {
+  static fluent.WidgetStateProperty<Color> colorToTextButtonState(Color color,
+      {bool inverted = false}) {
     final baseColor = inverted ? Colors.white : color;
-    
+
     return fluent.WidgetStateProperty.resolveWith((states) {
       if (states.isDisabled) {
         return baseColor.withAlpha((255 * 0.5).round());
@@ -59,14 +57,15 @@ class FluentColorExtensions {
       return baseColor;
     });
   }
-  
+
   /// Converts a Fluent UI [AccentColor] to a Material [MaterialColor].
-  /// 
+  ///
   /// This method creates a Material color swatch based on the Fluent accent color.
-  static MaterialColor accentColorToMaterialColor(fluent.AccentColor accentColor) {
+  static MaterialColor accentColorToMaterialColor(
+      fluent.AccentColor accentColor) {
     final color = accentColor.normal;
-    
-    return MaterialColor(color.value, {
+
+    return MaterialColor(color.value.toInt(), {
       50: _adjustBrightness(color, 0.4),
       100: _adjustBrightness(color, 0.3),
       200: _adjustBrightness(color, 0.2),
@@ -79,17 +78,17 @@ class FluentColorExtensions {
       900: _adjustBrightness(color, -0.3),
     });
   }
-  
+
   /// Adjusts the brightness of a color by the given factor.
-  /// 
+  ///
   /// A positive factor makes the color lighter, while a negative factor makes it darker.
   /// The factor should be between -1.0 and 1.0.
   static Color _adjustBrightness(Color color, double factor) {
     assert(factor >= -1.0 && factor <= 1.0);
-    
+
     final hsl = HSLColor.fromColor(color);
     final lightness = (hsl.lightness + factor).clamp(0.0, 1.0);
-    
+
     return hsl.withLightness(lightness).toColor();
   }
 }
